pyproject.toml
[tool.poetry]
name = "m2-hw13"
version = "0.1.0"
description = ""
authors = ["Pelmenoff <Titanfall@ukr.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
pydantic-settings = "^2.1.0"
uvicorn = "^0.24.0.post1"
sqlalchemy = "^2.0.23"
fastapi-mail = "^1.4.1"
python-jose = "^3.3.0"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
passlib = "^1.7.4"
slowapi = "^0.1.8"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
alembic = "^1.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
